refactor: 가격 입력 컴포넌트 타입 수정
