+++ src/app/_component/common/InputPrice/formatPrice.ts
